Rename "convert-mocha-tests" to just "convert-tests"
