+++ lib/cli.js
@@ -112,12 +112,12 @@ program
   });
 
 program
-  .command('convert-mocha-tests-to-async-await [testsPath]')
-  .description('Convert mocha tests to use async/await')
+  .command('convert-tests-to-async-await [testsPath]')
+  .description('Convert tests to use async/await')
   .option('--dry-run', 'Don\'t modify files, output JSON instead')
   .action(function(testsPath, options) {
       testsPath = testsPath || 'tests';
-      watson.transformMochaToAsync(testsPath, options.dryRun);
+      watson.transformTestsToAsyncAwait(testsPath, options.dryRun);
   });
 
 program
