+++ lib/commands/convert-mocha-tests-to-async-await.js
@@ -4,8 +4,8 @@ var Watson = require('../../index');
 var watson = new Watson();
 
 module.exports = {
-  name: 'watson:convert-mocha-tests-to-async-await',
-  description: 'Convert mocha tests to use async/await',
+  name: 'watson:convert-tests-to-async-await',
+  description: 'Convert tests to use async/await',
   works: 'insideProject',
   anonymousOptions: [
     '<path>'
@@ -15,6 +15,6 @@ module.exports = {
   ],
   run: function(commandOptions, rawArgs) {
     var path = rawArgs[0] ||  'tests';
-    watson.transformMochaToAsync(path, commandOptions.dryRun);
+    watson.transformTestsToAsyncAwait(path, commandOptions.dryRun);
   }
 };
