+++ lib/commands/index.js
@@ -13,5 +13,5 @@ module.exports = {
   'watson:remove-ember-k': require('./remove-ember-k'),
   'watson:replace-needs-with-injection': require('./replace-needs-with-injection'),
   'watson:remove-ember-data-is-new-serializer-api': require('./remove-ember-data-is-new-serializer-api'),
-  'watson:convert-mocha-tests-to-async-await': require('./convert-mocha-tests-to-async-await')
+  'watson:convert-mocha-tests-to-async-await': require('./convert-tests-to-async-await')
 };
