+++ lib/watson.js
@@ -16,7 +16,7 @@ var transformDestroyApp = require('./formulas/destroy-app-transform');
 var removeEmberK = require('./formulas/remove-ember-k');
 var replaceNeedsWithInjection = require('./formulas/replace-needs-with-injection');
 var removeEmberDataIsNewSerializerAPI = require('./formulas/remove-ember-data-is-new-serializer-api');
-var mochaAsync = require('./formulas/convert-mocha-tests-to-async-await');
+var transformTestsToAsyncAwait = require('./formulas/convert-tests-to-async-await');
 
 module.exports = EmberWatson;
 
@@ -115,14 +115,14 @@ EmberWatson.prototype.findOverloadedCPs = function(rootPath, dryRun) {
   return searcher;
 };
 
-EmberWatson.prototype._transformMochaToAsync = mochaAsync;
+EmberWatson.prototype._transformTestsToAsyncAwait = transformTestsToAsyncAwait;
 
-EmberWatson.prototype.transformMochaToAsync = function(rootPath, dryRun) {
+EmberWatson.prototype.transformTestsToAsyncAwait = function(rootPath, dryRun) {
     var tests  = findFiles(rootPath, '.js').filter(function(file){
         return file.indexOf('-test.js') > 0;
     });
 
-    transform(tests, this._transformMochaToAsync, dryRun);
+    transform(tests, this._transformTestsToAsyncAwait, dryRun);
 };
 
 function transform(files, transformFormula, dryRun) {
