+++ tests/convert-tests-to-async-await-test.js
@@ -5,7 +5,7 @@ var fs          = require('fs');
 var astEquality = require('./helpers/ast-equality');
 
 describe('Mocha async transformation', function() {
-  var baseDir = './tests/fixtures/convert-mocha-tests-to-async-await';
+  var baseDir = './tests/fixtures/convert-tests-to-async-await';
   var watson;
 
   beforeEach(function() {
@@ -14,7 +14,7 @@ describe('Mocha async transformation', function() {
 
   it('transforms async acceptance test using `done()` callback', function() {
     var source            = fs.readFileSync(baseDir + '/old-acceptance-test-with-done.js');
-    var newSource         = watson._transformMochaToAsync(source);
+    var newSource         = watson._transformTestsToAsyncAwait(source);
     var expectedNewSource = fs.readFileSync(baseDir + '/new-acceptance-test-with-done.js');
 
     astEquality(newSource, expectedNewSource);
@@ -22,7 +22,7 @@ describe('Mocha async transformation', function() {
 
   it('transforms async acceptance test without `done()`', function() {
     var source            = fs.readFileSync(baseDir + '/old-acceptance-test-without-done.js');
-    var newSource         = watson._transformMochaToAsync(source);
+    var newSource         = watson._transformTestsToAsyncAwait(source);
     var expectedNewSource = fs.readFileSync(baseDir + '/new-acceptance-test-without-done.js');
 
     astEquality(newSource, expectedNewSource);
@@ -30,7 +30,7 @@ describe('Mocha async transformation', function() {
 
   it('transforms async acceptance test nested `andThen`s', function() {
     var source            = fs.readFileSync(baseDir + '/old-acceptance-test-with-nested-andthen.js');
-    var newSource         = watson._transformMochaToAsync(source);
+    var newSource         = watson._transformTestsToAsyncAwait(source);
     var expectedNewSource = fs.readFileSync(baseDir + '/new-acceptance-test-with-nested-andthen.js');
 
     astEquality(newSource, expectedNewSource);
@@ -38,7 +38,7 @@ describe('Mocha async transformation', function() {
 
   it('leaves synchronous tests as-is', function() {
     var source            = fs.readFileSync(baseDir + '/acceptance-test-synchronous.js');
-    var newSource         = watson._transformMochaToAsync(source);
+    var newSource         = watson._transformTestsToAsyncAwait(source);
     var expectedNewSource = fs.readFileSync(baseDir + '/acceptance-test-synchronous.js');
 
     astEquality(newSource, expectedNewSource);
@@ -46,7 +46,7 @@ describe('Mocha async transformation', function() {
 
   it('leaves `done` callback as-is when used outside of `andThen`', function() {
     var source            = fs.readFileSync(baseDir + '/old-acceptance-test-with-callback.js');
-    var newSource         = watson._transformMochaToAsync(source);
+    var newSource         = watson._transformTestsToAsyncAwait(source);
     var expectedNewSource = fs.readFileSync(baseDir + '/new-acceptance-test-with-callback.js');
 
     astEquality(newSource, expectedNewSource);
@@ -54,7 +54,7 @@ describe('Mocha async transformation', function() {
 
   it('transforms async qunit test', function() {
     var source            = fs.readFileSync(baseDir + '/old-qunit.js');
-    var newSource         = watson._transformMochaToAsync(source);
+    var newSource         = watson._transformTestsToAsyncAwait(source);
     var expectedNewSource = fs.readFileSync(baseDir + '/new-qunit.js');
 
     astEquality(newSource, expectedNewSource);
