+++ tests/fixtures/convert-tests-to-async-await/acceptance-test-synchronous.js
